introduced ActorSpecBase to be used for testing actors
