+++ src/test/scala/castalia/actors/ManagerSpec.scala
@@ -1,30 +1,17 @@
-package castalia
+package castalia.actors
 
 import akka.actor.ActorSystem
 import akka.http.scaladsl.model.StatusCodes._
-import akka.testkit.{ImplicitSender, TestKit, TestProbe}
-import akka.util.Timeout
+import akka.testkit.TestProbe
 import castalia.management.Manager
 import castalia.model.Messages.{Done, UpsertEndpoint}
 import castalia.model.Model.{ResponseConfig, StubConfig}
-import org.scalatest._
 
-import scala.concurrent.duration._
 import scala.language.postfixOps
 
-class ManagerSpec(_system: ActorSystem) extends TestKit(_system)
-  with ImplicitSender
-  with WordSpecLike
-  with Matchers
-  with BeforeAndAfterAll {
+class ManagerSpec(_system: ActorSystem) extends ActorSpecBase(_system) {
 
   def this() = this(ActorSystem("ManagerSpec"))
-
-  override def afterAll {
-    TestKit.shutdownActorSystem(system)
-  }
-
-  private implicit val timeout = Timeout(2 seconds)
 
   "Manager actor" must {
